refactor(restapi): move search-stops reference builders into reference_utils.go
Relocate routeReferencesForStops/agencyReferencesForStops out of
search_stops_handler.go and into the shared reference-building file,
per CONTRIBUTING.md's Code Reuse convention. agencyReferencesForStops
now delegates to models.AgencyReferenceFromDatabase instead of
duplicating its null-handling.
